fix(pgproto): use u64 instead of u32 for portal & stmt storage ids

+++ src/pgproto/client.rs
@@ -16,7 +16,8 @@ mod startup;
 
 pub use startup::ClientParams;
 
-pub type ClientId = u32;
+/// We generate those sequentially for every client connection.
+pub type ClientId = u64;
 
 /// Postgres client representation.
 pub struct PgClient<S> {
